fix: Gemini 리뷰 코멘트 대응 (트리 검증 활성화, FSD 위반 수정 등)
