+++ src/features/question-ab/model/types.ts
@@ -1,10 +1,7 @@
-export type AbRatio = "9:16" | "1:1" | "4:3";
+import type { AbRatio } from "@/shared/constants/imageRatio";
 
-export const RATIO_TO_CSS: Record<AbRatio, string> = {
-  "9:16": "9/16",
-  "1:1": "1/1",
-  "4:3": "4/3",
-};
+export type { AbRatio };
+export { RATIO_TO_CSS } from "@/shared/constants/imageRatio";
 
 export interface AbQuestionData {
   title: string;
