+++ src/features/question-fivesec/answer/FivesecCountdownPhase.tsx
@@ -1,6 +1,6 @@
 import { motion } from "framer-motion";
 import { QuestionHeader } from "@/features/test-participate/ui/QuestionHeader";
-import { RATIO_TO_CSS } from "@/features/question-ab/model/types";
+import { RATIO_TO_CSS } from "@/shared/constants/imageRatio";
 
 interface Props {
   remaining: number;
